// backend/src/common/enums/notification-type.enum.ts
export enum NotificationType {
  BidAccepted = 'bid_accepted',
  ContractSigned = 'contract_signed',
  ContractCompleted = 'contract_completed',
  ContractTerminated = 'contract_terminated'
}

export enum NotificationRelatedType {
  Requirement = 'requirement',
  Contract = 'contract',
  Bid = 'bid'
}